added output to file + few changes

+++ Player.cpp
@@ -6,8 +6,8 @@
 
 //////// CLASS: MazeCell ////////
 
-MazeCell::MazeCell() : numOfLocsTried(0), obstacle(MazeObstacle::UNKNOWN){
-	for(int i = 0; i < 4; i++) {
+MazeCell::MazeCell() : numOfLocsTried(0), obstacle(MazeObstacle::UNKNOWN) {
+	for (int i = 0; i < 4; i++) {
 		triedDirection[i] = false;
 	}
 
@@ -18,7 +18,7 @@ MazeCell::~MazeCell() {}
 MazeCell::MazeCell(const MazeCell &copied) {
 	numOfLocsTried = copied.numOfLocsTried;
 	obstacle = copied.obstacle;
-	for(int i = 0; i < 4; i++) {
+	for (int i = 0; i < 4; i++) {
 		triedDirection[i] = copied.triedDirection[i];
 	}
 }
@@ -26,9 +26,9 @@ MazeCell::MazeCell(const MazeCell &copied) {
 
 //////// CLASS: Player ////////
 
-Player::Player() : widthKnown(false), heightKnown(false), placedBookmark(false), directionChosen(false), mappedMaze(){
+Player::Player() : widthKnown(false), heightKnown(false), placedBookmark(false), directionChosen(false), mappedMaze() {
 	//set seed for random
-	srand(time(NULL));
+	srand((int)time(NULL));
 	//add the starting cell to the memorized map
 	std::vector<MazeCell> firstRow;
 	MazeCell firstCell;
@@ -64,87 +64,87 @@ void Player::hitBookmark() {
 	int &curRow = curLocation[0];
 	int &curCol = curLocation[1];
 
-	if(!directionChosen) return;
+	if (!directionChosen) return;
 	//choose another direction
 	directionChosen = false;
 
-	if(lastDirection == Direction::UP || lastDirection == Direction::DOWN) {
-		if(heightKnown) return;
+	if (lastDirection == Direction::UP || lastDirection == Direction::DOWN) {
+		if (heightKnown) return;
 		heightKnown = true;
 		int tempRows = rows;
 		std::cout << "=======" << "curRow = " << curRow << std::endl;
 		rows = std::abs(curRow - bookmarkRow) - 1;
 		std::cout << "(*) rows = " << rows << std::endl;
 		int rowsToErase = tempRows - rows;
-		if(rowsToErase == 0) rowsToErase++;
+		if (rowsToErase == 0) rowsToErase++;
 		std::cout << "(**) rowsToErase = " << rowsToErase << std::endl;
 		//fix coordinates
 		std::cout << "(***) bookmarkRow = " << bookmarkRow << std::endl;
-		if(lastDirection == Direction::UP) {
-			bookmarkRow = bookmarkRow-rowsToErase;
+		if (lastDirection == Direction::UP) {
+			bookmarkRow = bookmarkRow - rowsToErase;
 			curRow = bookmarkRow;
 			std::cout << "bookmarkRow = " << bookmarkRow << std::endl;
 			//remove the $rowsToErase highest rows
-			for(int i = rowsToErase; i > 0; i--) {
-				for(int j = 0; j <= cols; j++) {
-					if(i < rowsToErase && mappedMaze.at(0).at(j).obstacle != MazeObstacle::UNKNOWN) {
-						mappedMaze.at(mappedMaze.size()-i).at(j) = mappedMaze.at(0).at(j);
+			for (int i = rowsToErase; i > 0; i--) {
+				for (int j = 0; j <= cols; j++) {
+					if (i < rowsToErase && mappedMaze.at(0).at(j).obstacle != MazeObstacle::UNKNOWN) {
+						mappedMaze.at(mappedMaze.size() - i).at(j) = mappedMaze.at(0).at(j);
 					}
 				}
 				mappedMaze.erase(mappedMaze.begin());
 			}
 		}
 		else { //Direction::DOWN
-			if(bookmarkRow > rows) bookmarkRow -= rowsToErase;
+			if (bookmarkRow > rows) bookmarkRow -= rowsToErase;
 			curRow = bookmarkRow;
 			std::cout << "bookmarkRow = " << bookmarkRow << std::endl;
 			//remove the $rowsToErase lowest rows
-			for(int i = rowsToErase; i > 0; i--) {
-				for(int j = 0; j <= cols; j++) {
-					if(i < rowsToErase && mappedMaze.at(mappedMaze.size()-1).at(j).obstacle != MazeObstacle::UNKNOWN) {
-						mappedMaze.at(i-1).at(j) = mappedMaze.at(mappedMaze.size()-1).at(j);
+			for (int i = rowsToErase; i > 0; i--) {
+				for (int j = 0; j <= cols; j++) {
+					if (i < rowsToErase && mappedMaze.at(mappedMaze.size() - 1).at(j).obstacle != MazeObstacle::UNKNOWN) {
+						mappedMaze.at(i - 1).at(j) = mappedMaze.at(mappedMaze.size() - 1).at(j);
 					}
 				}
-				mappedMaze.erase(mappedMaze.end()-1);
+				mappedMaze.erase(mappedMaze.end() - 1);
 			}
 		}
 	}
 	else {
-		if(widthKnown) return;
+		if (widthKnown) return;
 		widthKnown = true;
 		int tempCols = cols;
 		cols = std::abs(curCol - bookmarkCol) - 1;
 		std::cout << "(*) cols = " << cols << std::endl;
 		int colsToErase = tempCols - cols;
-		if(colsToErase == 0) colsToErase++;
+		if (colsToErase == 0) colsToErase++;
 		std::cout << "(**) colsToErase = " << colsToErase << std::endl;
-		if(lastDirection == Direction::LEFT) {
-			bookmarkCol = bookmarkCol-colsToErase;
+		if (lastDirection == Direction::LEFT) {
+			bookmarkCol = bookmarkCol - colsToErase;
 			std::cout << "(***) bookmarkCol = " << bookmarkCol << std::endl;
 			curCol = bookmarkCol;
 			//remove the $colsToErase leftmost columns
-			for(int j = 0; j <= rows; j++) {
+			for (int j = 0; j <= rows; j++) {
 				std::vector<MazeCell> &iRow = mappedMaze.at(j);
-				for(int i = colsToErase; i > 0 ; i--) {
-					if(i < colsToErase && iRow.at(0).obstacle != MazeObstacle::UNKNOWN) {
-						iRow.at(iRow.size()-i) = iRow.at(0);
+				for (int i = colsToErase; i > 0; i--) {
+					if (i < colsToErase && iRow.at(0).obstacle != MazeObstacle::UNKNOWN) {
+						iRow.at(iRow.size() - i) = iRow.at(0);
 					}
 					iRow.erase(iRow.begin());
 				}
 			}
 		}
 		else { //Direction::RIGHT
-			if(bookmarkCol > cols) bookmarkCol -= colsToErase;
+			if (bookmarkCol > cols) bookmarkCol -= colsToErase;
 			std::cout << "(***) bookmarkCol = " << bookmarkCol << std::endl;
 			curCol = bookmarkCol;
 			//remove the $colsToErase rightmost columns
-			for(int j = 0; j <= rows; j++) {
+			for (int j = 0; j <= rows; j++) {
 				std::vector<MazeCell> &iRow = mappedMaze.at(j);
-				for(int i = colsToErase; i > 0; i--) {
-					if(i < colsToErase && iRow.at(iRow.size()-1).obstacle != MazeObstacle::UNKNOWN) {
-						iRow.at(i-1) = iRow.at(iRow.size()-1);
+				for (int i = colsToErase; i > 0; i--) {
+					if (i < colsToErase && iRow.at(iRow.size() - 1).obstacle != MazeObstacle::UNKNOWN) {
+						iRow.at(i - 1) = iRow.at(iRow.size() - 1);
 					}
-					iRow.erase(iRow.end()-1);
+					iRow.erase(iRow.end() - 1);
 				}
 			}
 		}
@@ -166,7 +166,7 @@ void Player::hitWall() {
 	std::cout << "(****) lastDir = " << lastDir << std::endl;
 	undoMove(lastDir);
 	path.pop();
-	if(curLocation[0] != bookmarkLoc[0] || curLocation[1] != bookmarkLoc[1]) { //check if the bookmark is at the same location as the player, and place it if it isn't
+	if (curLocation[0] != bookmarkLoc[0] || curLocation[1] != bookmarkLoc[1]) { //check if the bookmark is at the same location as the player, and place it if it isn't
 		placedBookmark = false;
 	}
 }
@@ -231,31 +231,31 @@ Direction Player::move() {
 	MazeCell &cell = mappedMaze[curLocation[0]][curLocation[1]];
 	////updateAvailableDirections(curLocation[0], curLocation[1]);
 	// change current cell.obstacle to SPACE, as the player is on it
-	if(cell.obstacle == MazeObstacle::UNKNOWN) {
+	if (cell.obstacle == MazeObstacle::UNKNOWN) {
 		cell.obstacle = MazeObstacle::SPACE;
 	}
 
 	Direction nextDirection;
 	std::cout << "move(): 0" << std::endl;
-	std::string lDir = lastDirection == Direction::RIGHT? "RIGHT" : (lastDirection == Direction::LEFT ? "LEFT" : (lastDirection == Direction::UP ? "UP" : (lastDirection == Direction::DOWN? "DOWN" : "BOOKMARK")));
+	std::string lDir = lastDirection == Direction::RIGHT ? "RIGHT" : (lastDirection == Direction::LEFT ? "LEFT" : (lastDirection == Direction::UP ? "UP" : (lastDirection == Direction::DOWN ? "DOWN" : "BOOKMARK")));
 	std::cout << "lastDirection = " << lDir << std::endl;
 
-	if(!path.empty() && path.top() != Direction::BOOKMARK) {
-		switch(path.top()) {
+	if (!path.empty() && path.top() != Direction::BOOKMARK) {
+		switch (path.top()) {
 		case Direction::LEFT:
-			if(!cell.triedDirection[1]) cell.numOfLocsTried++;
+			if (!cell.triedDirection[1]) cell.numOfLocsTried++;
 			cell.triedDirection[1] = true; //Direction::RIGHT
 			break;
 		case Direction::RIGHT:
-			if(!cell.triedDirection[0]) cell.numOfLocsTried++;
+			if (!cell.triedDirection[0]) cell.numOfLocsTried++;
 			cell.triedDirection[0] = true; //Direction::LEFT
 			break;
 		case Direction::UP:
-			if(!cell.triedDirection[3]) cell.numOfLocsTried++;
+			if (!cell.triedDirection[3]) cell.numOfLocsTried++;
 			cell.triedDirection[3] = true; //Direction::DOWN
 			break;
 		case Direction::DOWN:
-			if(!cell.triedDirection[2]) cell.numOfLocsTried++;
+			if (!cell.triedDirection[2]) cell.numOfLocsTried++;
 			cell.triedDirection[2] = true; //Direction::UP
 			break;
 		default:
@@ -264,7 +264,7 @@ Direction Player::move() {
 
 	}
 	//place a bookmark if the player is not at the same cell as the bookmark, he haven't placed one and he isn't backtracking
-	if((bookmarkLoc[0] != curLocation[0] || bookmarkLoc[1] != curLocation[1]) && !placedBookmark && cell.numOfLocsTried < 4) {
+	if ((bookmarkLoc[0] != curLocation[0] || bookmarkLoc[1] != curLocation[1]) && !placedBookmark && cell.numOfLocsTried < 4) {
 		std::cout << "move(): 1" << std::endl;
 		placedBookmark = true;
 		bookmarkLoc[0] = curLocation[0];
@@ -273,26 +273,26 @@ Direction Player::move() {
 	}
 	else {
 		int chosenDir = 0;
-		if(directionChosen) { //continue along the same direction
+		if (directionChosen) { //continue along the same direction
 			nextDirection = path.top();
 			chosenDir = nextDirection;
-			if(!cell.triedDirection[chosenDir]){
+			if (!cell.triedDirection[chosenDir]) {
 				cell.triedDirection[chosenDir] = true;
 				cell.numOfLocsTried++;
 			}
 		}
-		else if(cell.numOfLocsTried < 4) { //choose a random direction that wasn't tried yet
+		else if (cell.numOfLocsTried < 4) { //choose a random direction that wasn't tried yet
 			////std::cout << "move(): 1" << std::endl;
-			int numAvailable = 4-cell.numOfLocsTried;
+			int numAvailable = 4 - cell.numOfLocsTried;
 			std::cout << "numAvailable = " << numAvailable << std::endl;
-			int randNum =rand()%numAvailable;
-			for(; randNum >= 0; chosenDir++) {
-				if(!cell.triedDirection[chosenDir]) randNum--;
+			int randNum = rand() % numAvailable;
+			for (; randNum >= 0; chosenDir++) {
+				if (!cell.triedDirection[chosenDir]) randNum--;
 			}
 			chosenDir--;
 			std::cout << "chosenDir = " << chosenDir << std::endl;
 			nextDirection = Direction(chosenDir);
-			if(cell.triedDirection[chosenDir]) {
+			if (cell.triedDirection[chosenDir]) {
 				std::cout << "ALERT ALERT ALERT ALERT ALERT ALERT ALERT ALERT ALERT ALERT ALERT ALERT ALERT ALERT" << std::endl;
 			}
 			////std::cout << "move(): 5" << std::endl;
@@ -304,7 +304,7 @@ Direction Player::move() {
 			std::cout << "!!!!!!backtracking!!!!!!" << std::endl;
 			directionChosen = false;
 			placedBookmark = false;
-			switch(path.top()) {
+			switch (path.top()) {
 			case Direction::LEFT:
 				nextDirection = Direction::RIGHT;
 				break;
@@ -323,10 +323,10 @@ Direction Player::move() {
 			path.pop();
 		}
 	}
-	if(nextDirection != Direction::BOOKMARK) {
+	if (nextDirection != Direction::BOOKMARK) {
 		updateLocation(nextDirection);
-		if(directionChosen) {
-			std::string nDir = nextDirection == Direction::RIGHT? "RIGHT" : (nextDirection == Direction::LEFT ? "LEFT" : (nextDirection == Direction::UP ? "UP" : (nextDirection == Direction::DOWN? "DOWN" : "BOOKMARK")));
+		if (directionChosen) {
+			std::string nDir = nextDirection == Direction::RIGHT ? "RIGHT" : (nextDirection == Direction::LEFT ? "LEFT" : (nextDirection == Direction::UP ? "UP" : (nextDirection == Direction::DOWN ? "DOWN" : "BOOKMARK")));
 			path.push(nextDirection);
 			updateMap();
 		}
@@ -341,7 +341,7 @@ Direction Player::move() {
  *
  */
 void Player::undoMove(Direction dir) {
-	switch(dir) {
+	switch (dir) {
 	case Direction::LEFT:
 		updateLocation(Direction::RIGHT);
 		break;
@@ -362,14 +362,14 @@ void Player::undoMove(Direction dir) {
 /**
  * Updates the player's location.
  */
-void Player::updateLocation(Direction dir){
+void Player::updateLocation(Direction dir) {
 	int &rows = knownDimensions[0];
 	int &cols = knownDimensions[1];
 	int &curRow = curLocation[0];
 	int &curCol = curLocation[1];
-	switch(dir) {
+	switch (dir) {
 	case Direction::UP:
-		if(heightKnown && curRow == 0) {
+		if (heightKnown && curRow == 0) {
 			curRow = rows;
 		}
 		else {
@@ -377,7 +377,7 @@ void Player::updateLocation(Direction dir){
 		}
 		break;
 	case Direction::DOWN:
-		if(heightKnown && curRow == rows) {
+		if (heightKnown && curRow == rows) {
 			curRow = 0;
 		}
 		else {
@@ -385,7 +385,7 @@ void Player::updateLocation(Direction dir){
 		}
 		break;
 	case Direction::LEFT:
-		if(widthKnown && curCol == 0) {
+		if (widthKnown && curCol == 0) {
 			curCol = cols;
 		}
 		else {
@@ -393,7 +393,7 @@ void Player::updateLocation(Direction dir){
 		}
 		break;
 	case Direction::RIGHT:
-		if(widthKnown && curCol == cols) {
+		if (widthKnown && curCol == cols) {
 			curCol = 0;
 		}
 		else {
@@ -412,10 +412,10 @@ void Player::updateMap() {
 	int &curCol = curLocation[1];
 	std::vector<MazeCell> newRow;
 	// the player went lower than previously known
-	if(curRow > rows) {
+	if (curRow > rows) {
 		rows++;
 		// add the number of known columns to the row.
-		for(int i = 0; i <= cols; i++) {
+		for (int i = 0; i <= cols; i++) {
 			newRow.push_back(MazeCell());
 		}
 		//insert row at the end (the lowest row)
@@ -423,13 +423,13 @@ void Player::updateMap() {
 	}
 
 	// the player went higher than previously known
-	else if(curRow < 0) {
+	else if (curRow < 0) {
 		rows++;
 		//adjust location so there won't be any negative coordinations
 		curRow++;
 		bookmarkLoc[0]++;
 		// add the number of known columns to the row
-		for(int i = 0; i <= cols; i++) {
+		for (int i = 0; i <= cols; i++) {
 			newRow.push_back(MazeCell());
 		}
 		//insert row at the beginning (the highest row)
@@ -437,22 +437,22 @@ void Player::updateMap() {
 	}
 
 	// the player went more right than previously known
-	else if(curCol > cols) { // column needs to be added
+	else if (curCol > cols) { // column needs to be added
 		cols++;
 		// add a cell at the end of each row
-		for(int i = 0; i <= rows; i++) {
+		for (int i = 0; i <= rows; i++) {
 			//insert cell at the end of the row (the rightmost cell)
 			mappedMaze[i].push_back(MazeCell());
 		}
 	}
 
 	// the player went more left than previously known
-	else if(curCol < 0) {
+	else if (curCol < 0) {
 		cols++;
 		//adjust location so there won't be any negative coordinations
 		curCol++;
 		bookmarkLoc[1]++;
-		for(int i = 0; i <= rows; i++) {
+		for (int i = 0; i <= rows; i++) {
 
 			std::vector<MazeCell> &iRow = mappedMaze.at(i);
 			//insert cell at the beginning of the row (the leftmost cell)
@@ -470,19 +470,19 @@ void Player::printMaze() {
 	int &curCol = curLocation[1];
 	std::cout << "\t\trows = " << rows << ", cols = " << cols << std::endl;
 	std::cout << "\t\tcurRow = " << curRow << ", curCol = " << curCol << std::endl;
-	for(int i = 0; i <= rows; i++) {
+	for (int i = 0; i <= rows; i++) {
 		std::cout << "\t\t";
-		for(int j = 0; j <= cols; j++) {
+		for (int j = 0; j <= cols; j++) {
 			MazeCell cell = mappedMaze[i][j];
-			if(i == curRow && j == curCol) {
+			if (i == curRow && j == curCol) {
 				std::cout << "@";
 				continue;
 			}
-			if(i == bookmarkLoc[0] && j == bookmarkLoc[1]) {
+			if (i == bookmarkLoc[0] && j == bookmarkLoc[1]) {
 				std::cout << "B";
 				continue;
 			}
-			switch(cell.obstacle) {
+			switch (cell.obstacle) {
 			case MazeObstacle::SPACE:
 				std::cout << "_";
 				break;
